+++ client/src/App.tsx
@@ -2,8 +2,8 @@ import { Route, Switch } from 'react-router-dom';
 import GlobalStyles from '@style/GlobalStyles';
 
 import Header from '@components/Header/Header';
-import Main from '@page/Main';
-import Confirm from '@page/Confirm';
+import Main from '@page/Main/Main';
+import Confirm from '@page/Confirm/Confirm';
 
 export default function App() {
   return (
